% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angleCalc.R
\name{angleCalc}
\alias{angleCalc}
\title{Calculate angle between two sf point objects}
\usage{
angleCalc(x, y)
}
\arguments{
\item{x}{numeric vector with two elements, X and Y coordinates of a point}

\item{y}{numeric vector with two elements, X and Y coordinates of a point}
}
\value{
azimuthal from x to y, in degrees
}
\description{
Calculate angle between two sf point objects
}
\examples{
p1 <- c(0, 1)
p2 <- c(1, 2)
angleCalc(p1, p2)

}
