% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baLarger.R
\name{baLarger}
\alias{baLarger}
\title{Basal area of larger trees - Wykoff et al. 1982}
\usage{
baLarger(ba, focal_size, size)
}
\arguments{
\item{ba}{vector of basal area of competitor trees}

\item{focal_size}{size of subject tree}

\item{size}{vector of sizes of competitor trees}
}
\value{
value of competition index for focal tree
}
\description{
Basal area of larger trees - Wykoff et al. 1982
}
\details{
Returns the sum of basal areas of all trees larger than the focal
tree. In Wykoff et al. (1982), \code{size} is basal area, though other
measures of size could be used such as tree height (e.g. Flake et al. 2022).
}
\references{
Wykoff, W. R., Crookston, N. L., Stage, A. R. (1982). User's
guide to the stand prognosis model. United States Department of Agriculture
Forest Service. Ogden UT, USA.
}
