% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baLocal.R
\name{baLocal}
\alias{baLocal}
\title{Total basal area of competitor trees - Steneker and Jarvis 1963}
\usage{
baLocal(ba)
}
\arguments{
\item{ba}{vector of basal area of competitor trees}
}
\value{
atomic vector of competition index for focal tree
}
\description{
Total basal area of competitor trees - Steneker and Jarvis 1963
}
\details{
Included mainly for posterity, simply the sum of basal areas of all
competitor trees.
}
\references{
Steneker, G.A.; Jarvis, J.M. (1963). A preliminary study to
assess competition in a white spruce-trembling aspen stand. Forestry
Chronicle. Volume 39. Issue 3. Pages 334-336.
}
