% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basalArea.R
\name{basalArea}
\alias{basalArea}
\title{Calculate basal area from DBH}
\usage{
basalArea(x)
}
\arguments{
\item{x}{numeric vector of DBH values}
}
\value{
numeric vector
}
\description{
Calculate basal area from DBH
}
\examples{
a <- c(1.23, 5.67, 10.11)
basalArea(a)
}
