% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicuar.R
\docType{data}
\name{bicuar}
\alias{bicuar}
\title{Tree stem data from a 1 ha woodland plot in Bicuar National Park, Angola}
\format{
An object of class \code{data.frame} with 460 rows and 6 columns.
}
\usage{
data(bicuar)
}
\description{
Tree stem data from a 1 ha woodland plot in Bicuar National Park, Angola
}
\keyword{datasets}
