\name{BLI_2017}
\alias{BLI_2017}
\title{Better Life Index 2017 indicators}
\description{Data related to BLI Edition 2017 (OECD, 2017) for all 38 OECD and non-OECD countries (Data extracted on: 19\02\2020).

For more info, please see \url{https://data-explorer.oecd.org}.
}

\usage{data(BLI_2017)}

\format{BLI_2017 is a dataset with 38 observations and 12 indicators.

\describe{
	\item{country}{OECD and non-OECD countries.}	
    \item{housing}{Housing.}	
    \item{income}{Income and wealth.}	
    \item{jobs}{Jobs and earnings.}	
    \item{community}{Community engagement.}	
    \item{education}{Education.}	
    \item{environment}{Environment quality.}	
    \item{civic}{Civic engagement.}	
    \item{health}{Health.}	
    \item{satisfaction}{Life satisfaction.}	
    \item{safety}{Personal security (safety).}	
    \item{worklife}{Work-Life balance.}	
}}

\author{Fusco E.}
\examples{data(BLI_2017)}