% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarkEvans.R
\name{clarkEvans}
\alias{clarkEvans}
\title{Clark-Evans index of neighbourhood pattern}
\usage{
clarkEvans(x, area)
}
\arguments{
\item{x}{two column matrix of individual x and y coordinates}

\item{area}{area of the plot, in the same units as \code{x} and
\code{y} coordinates}
}
\value{
Value of the competition index for the structural unit, i.e. plot.
}
\description{
Clark-Evans index of neighbourhood pattern
}
\details{
Essentially the sum of nearest neighbour distances of all
individuals in the plot, normalised by the density of individuals in
the plot. Bounded between 0 and 2.15, with values <1 indicating a
clustered distribution. A completely regular hexagonal distribution
results in the highest value of 2.15.
}
\examples{
data(bicuar)
clarkEvans(bicuar[,c("x", "y")], 10000)

}
\references{
Clark, F. J., Evans, F. C. (1954). Distance to the nearest
neighbour as a measure of spatial relationships in populations. Ecology.
Volume 35. Pages 445-453
}
