% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compInd.R
\docType{package}
\name{compInd-package}
\alias{compInd}
\alias{compInd-package}
\title{compInd: Forest tree-tree competition indices}
\description{

}
\details{
A collection of functions drawn from the academic literature used
in forests to quantify tree-tree competition.
}
\author{
John L. Godlee (johngodlee@gmail.com)
}
