% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crowdInd.R
\name{crowdInd}
\alias{crowdInd}
\title{Forest matrix crowding index - Seydack et al. 2011}
\usage{
crowdInd(dbh, dbh_min = 10, dbh_max = 30)
}
\arguments{
\item{dbh}{vector of DBH (diameter at breast height) measurements of competitor trees}

\item{dbh_min}{minimum DBH threshold considered}

\item{dbh_max}{maximum DBH threshold considered}
}
\value{
value of competition index for focal tree
}
\description{
Forest matrix crowding index - Seydack et al. 2011
}
\details{
A simple competition index which returns the sum of DBH values of
all competitor trees within user-defined DBH thresholds. Armin et al.
(2011) use a minimum threshold of 10 cm DBH and a maximum threshold of
30 cm DBH.
}
\examples{
data(bicuar)
nb <- nearNeighb(bicuar$x, bicuar$y, bicuar$stem_id, k = 4) 
lapply(nb, function(x) {
  nb <- merge(x, bicuar, by.x = "nb", by.y = "stem_id")
  crowdInd(nb$diam, dbh_min = 10, dbh_max = 30)
  })

}
\references{
Armin H.W. Seydack, Graham Durrheim, Josua H. Louw.
Spatiotemporally interactive growth dynamics in selected South African
forests: Edaphoclimatic environment, crowding and climate effects.
Forest Ecology and Management. Volume 261. Issue 7. 2011.
Pages 1152-1169.
}
