% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGen.R
\name{dataGen}
\alias{dataGen}
\title{Generate fake data of a rectangular tree plot}
\usage{
dataGen(
  nplots = 5,
  min_stems = 200,
  max_stems = 500,
  min_diam = 5,
  max_diam = 100,
  plot_width = 100,
  plot_length = plot_width,
  species = LETTERS[1:20]
)
}
\arguments{
\item{nplots}{number of plots}

\item{min_stems}{minimum number of stems per plot}

\item{max_stems}{maximum number of stems per plot}

\item{min_diam}{minimum stem diameter}

\item{max_diam}{maximum stem diameter}

\item{plot_width}{plot width}

\item{plot_length}{plot length}

\item{species}{vector of species names from which to sample}
}
\value{
dataframe, where each row is a tree stem. Default of five plots,
each with between 200 and 500 stems, from 20 species, with stem diameter
values between 5 and 100. All trees have a single stem. Diameter values
are drawn from a uniform distribution. Stem locations and species are
randomly sampled, with repeats.
}
\description{
Generate fake data of a rectangular tree plot
}
\examples{
dat <- dataGen()
dat2 <- dataGen(nplots = 1, min_stems = 10, max_stems = 50, dbh_min = 10, 
  dbh_max = 200, plot_width = 20, plot_length = 50, 
  sp = c("Burkea africana", "Ochna pulchra"))

}
