% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhCorr.R
\name{dbhCorr}
\alias{dbhCorr}
\title{Diameter correlation index - Davies 2008}
\usage{
dbhCorr(dbh, focal_dbh)
}
\arguments{
\item{dbh}{vector of DBH (diameter at breast height) measurements of competitor trees}

\item{focal_dbh}{DBH of focal tree}
}
\value{
atomic vector of competition index for focal tree
}
\description{
Diameter correlation index - Davies 2008
}
\details{
All else being equal, the value of \code{diamCorr()} increases as
the reference tree size increases. For a given reference tree size, the
value increases as average neighbour size increases. Given by the equation:
\deqn{DCI_{i} = \frac{dbh_{i} \sum_{j=1}^{n} dbh_{j}}{n \overbar{dbh}^2}}
}
\references{
Davis, O., Pommerening, A. (2008). The contribution of
structural indices to the modelling of Sitka spruce (Picea sitchensis) and
birch (Betula spp.) crowns. Forest Ecology and Management. Volume 256.
Pages 68-77.
}
