% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhDiff.R
\name{dbhDiff}
\alias{dbhDiff}
\title{DBH differentiation - Pommerening 2002}
\usage{
dbhDiff(dbh, focal_dbh)
}
\arguments{
\item{dbh}{vector of dbh (diameter at breast height) measurements of competitor trees}

\item{focal_dbh}{dbh of focal tree}
}
\value{
atomic vector of competition index for focal tree
}
\description{
DBH differentiation - Pommerening 2002
}
\details{
Gives the size difference of neighbouring trees and describes the
spatial distribution of tree sizes. The value increases with increasing
average size difference between neighbuoring trees. \code{dbhDiff == 0}
when all neighbours have equal size.
}
\references{
Pommerening, A. (2002). Approaches to quantifying forest
structures. Forestry, Volume 75, Issue 3. Pages 305-324.
}
