% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhDom.R
\name{dbhDom}
\alias{dbhDom}
\title{DBH dominance - Aguirre et al. 2003}
\usage{
dbhDom(dbh, focal_dbh)
}
\arguments{
\item{dbh}{vector of dbh (diameter at breast height) measurements of
competitor trees}

\item{focal_dbh}{dbh of focal tree}
}
\value{
atomic vector of competition index for focal tree
}
\description{
DBH dominance - Aguirre et al. 2003
}
\details{
Gives the proportion of the n nearest neighbours which are smaller
than the focal tree.
}
\references{
Aguirre, O., Hui, G., von Gadow, K., Jimenez, J. (2003). An
analysis of spatial forest structure using neighbourhood-based variables.
Forest Ecology and Management. Volume 183. Pages 13.
}
