% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domConInd.R
\name{domConInd}
\alias{domConInd}
\title{Dominance concentration index (Su et al. 2020)}
\usage{
domConInd(ba, sp)
}
\arguments{
\item{ba}{vector of basal area measurements of competitor trees}

\item{sp}{vector of individual species names}
}
\value{
value of dominance concentration index for focal tree
}
\description{
Dominance concentration index (Su et al. 2020)
}
\examples{
data(bicuar)
nb <- nearNeighb(bicuar$x, bicuar$y, bicuar$stem_id, k = 10) 
lapply(nb, function(x) {
  nb <- merge(x, bicuar, by.x = "nb", by.y = "stem_id")
  nb$ba <- basalArea(nb$diam)
  domConInd(nb$ba, nb$sp)
  })

}
\references{
Su S., Guan B. T., Chang-Yang C., Sun I., Wang H., Hsieh C.
(2020). Multi-stemming and size enhance survival of dominant tree
species in a frequently typhoon-disturbed forest. Journal of Vegetation
Science 31(3), pp. 429-439. DOI: 10.1111/jvs.12858
}
