\name{EU_2020}
\alias{EU_2020}
\title{Europe 2020 indicators}
\description{Europe 2020, a strategy for jobs and smart, sustainable and inclusive growth, is based on five EU headline targets which are currently measured by eight headline indicators, Headline indicators, Eurostat, year 1990-2012 (Last update:	21/11/2013).

For more info, please see \url{https://ec.europa.eu/eurostat/en/web/products-statistics-in-focus/-/KS-SF-12-039}.
}
\usage{data(EU_2020)}

\format{EU_2020 is a dataset with 30 observations and 12 indicators (190 indicator per year).

\describe{
	\item{geo}{EU-Member States including EU (28 countries) and EU (27 countries) row.}	
    \item{employXXXX}{Employment rate - age group 20-64, year XXXX (1992-2012).}	
    \item{perc_GDPXXXX}{Gross domestic expenditure on R&D (GERD), year XXXX (1990-2012).}
    \item{gas_emissXXXX}{Greenhouse gas emissions - base year 1990, year XXXX (1990-2011).}
	\item{share_renXXXX}{Share of renewable energy in gross final energy consumption, year XXXX (2004-2011).}
	\item{prim_enerXXXX}{Primary energy consumption, year XXXX (1990-2011).}
	\item{final_energyXXXX}{Final energy consumption, year XXXX (1990-2011).}
	\item{final_energyXXXX}{Early leavers from education and training - Perc. of the population aged 18-24 with at most lower secondary education and not in further education or training, year XXXX (1992-2012).}	
	\item{tertiaryXXXX}{Tertiary educational attainment - age group 30-34, year XXXX (2000-2012).}
	\item{risk_povertyXXXX}{People at risk of poverty or social exclusion - 1000 persons Perc. of total population, year XXXX (2004-2012).}
	\item{low_workXXXX}{People living in households with very low work intensity - 1000 persons Perc. of total population, year XXXX (2004-2012).}
	\item{risk_povertyXXXX}{People at risk of poverty after social transfers - 1000 persons Perc. of total population, year XXXX (2003-2012).}
	\item{deprivedXXXX}{Severely materially deprived people - 1000 persons Perc. of total population, year XXXX (2003-2012).}	
}}

\references{\url{https://ec.europa.eu/info/business-economy-euro/economic-and-fiscal-policy-coordination/eu-economic-governance-monitoring-prevention-correction/european-semester_en}}
\author{Vidoli F.}
\examples{data(EU_2020)}