\name{data_HPI}
\alias{data_HPI}
\title{Happy Planet Index 2017-2019 indicators}
\description{Data related to Happy Planet Index for 151 countries and the period 2017-2019.

For more info, please see \url{https://happyplanetindex.org}.
}
\usage{data(data_HPI)}

\format{data_HPI is a dataset with 453 observations and 10 variables.

\describe{
	\item{Country}{Country name}	
    \item{ISO}{ISO code}	
    \item{year}{Years 2017-2019}	
    \item{Continent}{Continent}	
    \item{Population}{Population (thousands)}	
    \item{Life_Expectancy}{Life Expectancy (years)}	
    \item{Ladder_of_life}{Ladder of life (Wellbeing) (0-10)}	
    \item{Ecological_Footprint}{Ecological Footprint (g ha)}	
    \item{HPI}{HPI}	
    \item{GDP_per_capita}{GDP per capita ($)}	
}}

\references{\url{https://happyplanetindex.org}}
\author{Fusco E.}
\examples{data(data_HPI)}