% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorimerComp.R
\name{lorimerComp}
\alias{lorimerComp}
\title{Lorimer's competition index - Lorimer 1983}
\usage{
lorimerComp(dbh, dist, focal_dbh, czr)
}
\arguments{
\item{dbh}{vector of DBH (diameter at breast height) measurements of competitor trees}

\item{dist}{vector of distances from focal tree to competitor trees}

\item{focal_dbh}{DBH of focal tree}

\item{czr}{Competition zone radius (CZR), normally based on plot-level stem density}
}
\value{
atomic vector of competition index for focal tree
}
\description{
Lorimer's competition index - Lorimer 1983
}
\details{
Value increases with dbh of competitor trees, decreases as the
focal tree DBH increases. Value increases as the distance of competitor
trees decreases. Tree distances are divided by the CZR in order to account
for "stand age".
}
\references{
Lorimer, C. G. (1983). Tests of age-independent competition
indices for individual trees in natural hardwood stands. Forest Ecology and
Management. Volume 6. Pages 343-360.
}
\seealso{
\code{\link[=lorimerCZR]{lorimerCZR()}} to calculate the CZR
}
