% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martinEk.R
\name{martinEk}
\alias{martinEk}
\title{Martin and Ek 1984}
\usage{
martinEk(dbh, dist, focal_dbh)
}
\arguments{
\item{dbh}{vector of DBH (diameter at breast height) measurements of
competitor trees}

\item{dist}{vector of distances from focal tree to competitor trees}

\item{focal_dbh}{DBH of focal tree}
}
\value{
atomic vector of competition index for focal tree
}
\description{
Martin and Ek 1984
}
\details{
Gives the sum of ratio of competitor to focal tree DBHs, multiplied
by the exponential of distances divided by competitor DBH plus focal DBH.
Given by the equation:
\deqn{\sum_{j\neq{}i}^{n} (d_{j} / d_{i}) exp((16l_{ij}) / (d_{i} + d_{j}))}
}
\references{
Martin G. L., Ek A. R. (1984). A Comparison of Competition
Measures and Growth Models for Predicting Plantation Red Pine Diameter and
Height Growth. Forest Science. Volume 30. Pages 731-743.
}
