% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pielou.R
\name{pielou}
\alias{pielou}
\title{Pielou's index of non-randomness}
\usage{
pielou(x, y, xmin, xmax, ymin, ymax, k)
}
\arguments{
\item{x}{vector of individual x axis coordinates}

\item{y}{vector of individual y axis coordinates}

\item{xmin}{minimum x coordinate in plot}

\item{xmax}{maximum x coordinate in plot}

\item{ymin}{minimum y coordinate in plot}

\item{ymax}{maximum y coordinate in plot}

\item{k}{number of randomly allocated sample points}
}
\value{
value of the competition index for the structural unit, i.e. plot.
}
\description{
Pielou's index of non-randomness
}
\details{
The sum of squared nearest neighbour distances normalised by the number of sample points and the number of individuals in the structural unit. Defined by the equation: \deqn{\pi \frac{n}{A} \frac{1}{k} \sum_{1}^{k} r_{i}^{2}} where \eqn{n} is the number of individuals in the structural unit, \eqn{A} is the structural unit area, \eqn{k} is the number of sample points, and \eqn{r_{i}} is the nearest neighbour distance to individual \eqn{i}.

As the sample points are randomly allocated within the bounds of
xmin,xmax,ymin,ymax, the mean of a number of runs of this function
could be used to further constrain the estimate of Pielou's index.
}
\examples{
data(bicuar)
pielou(bicuar$x_grid, bicuar$y_grid, 0, 100, 0, 100, 4)

}
\references{
Pielou, E. C. (1959). The use of point to plant distances in
the study of the pattern of plan populations. Journal of Ecology.
Volume 47. Pages 607-613.
}
