% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointDens.R
\name{pointDens}
\alias{pointDens}
\title{Point density - Purr 1962}
\usage{
pointDens(dbh, dist)
}
\arguments{
\item{dbh}{vector of DBH (diameter at breast height) measurements of
competitor trees}

\item{dist}{vector of distances from focal tree to competitor trees}
}
\value{
atomic vector of competition index for focal tree
}
\description{
Point density - Purr 1962
}
\references{
Spurr, S. H. (1962). A measure of point density. Forest Science.
Volume 8. Issue 1. Pages 85–96.
}
