% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialMingling.R
\name{spatialMingling}
\alias{spatialMingling}
\title{von Gadow's spatial mingling index}
\usage{
spatialMingling(x, y, sp, k = 4, adj = FALSE)
}
\arguments{
\item{x}{vector of individual x axis coordinates}

\item{y}{vector of individual y axis coordinates}

\item{sp}{vector of individual species names}

\item{k}{number of neighbours to consider}

\item{adj}{logical, if TRUE the basic spatial mingling index is multiplied
by Si/nmax, where Si is the number of species in the neighbourhood of
the focal tree, and nmax is the maximum number of species possible in the
neighbourhood, including the focal tree, i.e. k + 1.}
}
\value{

}
\description{
von Gadow's spatial mingling index
}
\references{
von Gadow, K., Hui, G. Y. (2001). Characterising forest spatial
structure and diversity. Sustainable Forestry in Temperate Regions. Proc. of
an international workshop organized at the University of Lund, Sweden.
Pages 20-30.
}
