% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angleCalc.R
\name{angleCalc}
\alias{angleCalc}
\title{Calculate angle between two sf point objects}
\usage{
angleCalc(x, y)
}
\arguments{
\item{x}{point feature of class 'sf'}

\item{y}{point feature of class 'sf'}
}
\value{
azimuthal from x to y, in degrees
}
\description{
Calculate angle between two sf point objects
}
\examples{
p1 <- st_point(c(0,1))
p2 <- st_point(c(1,2))
angleCalc(p1, p2)

}
