% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baLarger.R
\name{baLarger}
\alias{baLarger}
\title{Basal area of larger trees - Wykoff et al. 1982}
\usage{
baLarger(ba, focal_ba)
}
\arguments{
\item{ba}{vector of basal area of competitor trees}

\item{focal_ba}{atomic vector of basal area of subject tree}
}
\value{
atomic vector of competition index for focal tree
}
\description{
Basal area of larger trees - Wykoff et al. 1982
}
\details{
Returns the sum of basal areas of all trees with basal area larger
than the focal tree.
}
\references{
Wykoff, W. R., Crookston, N. L., Stage, A. R. (1982). User's
guide to the stand prognosis model. United States Department of Agriculture
Forest Service. Ogden UT, USA.
}
