% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicuar.R
\docType{data}
\name{bicuar}
\alias{bicuar}
\title{Data from a 1 ha woodland plot in Bicucar National Park, Angola}
\format{
An object of class \code{data.frame} with 460 rows and 6 columns.
}
\usage{
data(bicuar)
}
\description{
Data from a 1 ha woodland plot in Bicucar National Park, Angola
}
\author{
John L. Godlee
}
\keyword{datasets}
