\name{Compind-package}
\alias{Compind-package}
%\alias{Compind}
\docType{package}

\title{Composite Indicators - Compind}

\description{Compind package contains functions to enhance several approaches to the Composite Indicators (CIs) methods, focusing, in particular, on the normalisation and weighting-aggregation steps.}


\author{
Francesco Vidoli, Elisa Fusco
Maintainer: Francesco Vidoli <fvidoli@gmail.com>
}


\references{

Daraio, C., Simar, L. (2005) "\emph{Introducing environmental variables in nonparametric frontier models: a probabilistic approach}", Journal of productivity analysis, 24(1), 93-121.

Fusco E. (2015) "\emph{Enhancing non compensatory composite indicators: A directional proposal}", European Journal of Operational Research, 242(2), 620-630.

Fusco E. (2023) "\emph{Potential improvements approach in composite indicators construction: the Multi-directional Benefit of the Doubt model}", Socio-Economic Planning Sciences, vol. 85, 101447

Fusco, E., Liborio, M.P., Rabiei-Dastjerdi, H., Vidoli, F., Brunsdon, C. and Ekel, P.I. (2023), Harnessing Spatial Heterogeneity in Composite Indicators through the Ordered Geographically Weighted Averaging (OGWA) Operator. Geographical Analysis. https://doi.org/10.1111/gean.12384

R. Lahdelma, P. Salminen (2001) "\emph{SMAA-2: Stochastic multicriteria acceptability analysis for group decision making}", Operations Research, 49(3), pp. 444-454

OECD (2008) "\emph{Handbook on constructing composite indicators: methodology and user guide}".

Mazziotta C., Mazziotta M., Pareto A., Vidoli F. (2010) "\emph{La sintesi di indicatori territoriali di dotazione infrastrutturale: metodi di costruzione e procedure di ponderazione a confronto}", Rivista di Economia e Statistica del territorio, n.1.

Melyn W. and Moesen W.W. (1991) "\emph{Towards a synthetic indicator of macroeconomic performance: unequal weighting when limited information is available}", Public Economic research Paper 17, CES, KU Leuven.

Van Puyenbroeck T. and Rogge N. (2017) "\emph{Geometric mean quantity index numbers with Benefit-of-the-Doubt weights}", European Journal of Operational Research, 256(3), 1004-1014.

Rogge N., de Jaeger S. and Lavigne C. (2017) "\emph{Waste Performance of NUTS 2-regions in the EU: A Conditional Directional Distance Benefit-of-the-Doubt Model}", Ecological Economics, vol.139, pp. 19-32.

Simar L., Vanhems A. (2012) "\emph{Probabilistic characterization of directional distances and their robust versions}", Journal of Econometrics, 166(2), 342-354.

UNESCO (1974)"\emph{Social indicators: problems of definition and of selection}", Paris.

Vidoli F., Fusco E., Mazziotta C. (2015) "\emph{Non-compensability in composite indicators: a robust directional frontier method}", Social Indicators Research, 122(3), 635-652.

Vidoli F., Mazziotta C. (2013) "\emph{Robust weighted composite indicators by means of frontier methods with an application to European infrastructure endowment}", Statistica Applicata, Italian Journal of Applied Statistics.

Zanella A., Camanho A.S. and Dias T.G. (2015) "\emph{Undesirable outputs and weighting schemes in composite indicators based on data envelopment analysis}", European Journal of Operational Research, vol. 245(2), pp. 517-530.

}