% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbhDiff.R
\name{dbhDiff}
\alias{dbhDiff}
\title{DBH differentiation - Pommerening 2002}
\usage{
dbhDiff(dbh, focal_dbh)
}
\arguments{
\item{dbh}{vector of dbh (diameter at breast height) measurements of competitor trees}

\item{focal_dbh}{dbh of focal tree}
}
\value{
value of competition index for focal tree
}
\description{
DBH differentiation - Pommerening 2002
}
\details{
Gives the size difference of neighbouring trees and describes the
spatial distribution of tree sizes. The value increases with increasing
average size difference between neighbouring trees. \code{dbhDiff == 0}
when all neighbours have equal size.
}
\examples{
data(bicuar)
nb <- nearNeighb(bicuar$x, bicuar$y, bicuar$stem_id, k = 4) 
lapply(nb, function(x) {
  nb <- merge(x, bicuar, by.x = "nb", by.y = "stem_id")
  focal_diam <- unique(bicuar[bicuar$stem_id == nb$focal,"diam"])
  dbhDiff(nb$diam, focal_diam)
  })

}
\references{
Pommerening, A. (2002). Approaches to quantifying forest
structures. Forestry, Volume 75, Issue 3. Pages 305-324.
}
