% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeExclude.R
\name{edgeExclude}
\alias{edgeExclude}
\title{Find individuals inside a buffer zone away from the plot edge}
\usage{
edgeExclude(x, buffer, xmin, xmax, ymin, ymax)
}
\arguments{
\item{x}{two column matrix of individual x and y coordinates}

\item{buffer}{size of buffer zone inside plot, same scale as coordinates
\code{xmin}, \code{xmax}, \code{ymin}, \code{ymax}}

\item{xmin}{minimum x coordinate in plot}

\item{xmax}{maximum x coordinate in plot}

\item{ymin}{minimum y coordinate in plot}

\item{ymax}{maximum y coordinate in plot}
}
\value{
vector of rows from \code{x} which are not within the buffer zone.
}
\description{
Find individuals inside a buffer zone away from the plot edge
}
\details{
Generally, the buffer size should be the same size as the expected
competition zone radius around each tree, to ensure that the value of a
given competition index isn't under-estimated for trees near the plot
edge, due to a lack of data collected outside the plot. For example, if
the competition radius set by \code{nearNeighb} is 5 m, the buffer
should also be 5 m. Trees within the buffer zone are generally excluded
as focal trees in competition indices, but may still be used as
competitor trees by other focal trees not inside the buffer zone.
}
\examples{
data(bicuar)
edgeExclude(bicuar[,c("x", "y")], 5, 0, 100, 0, 100)

}
