% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martinEk.R
\name{martinEk}
\alias{martinEk}
\title{Martin and Ek 1984}
\usage{
martinEk(dbh, dist, focal_dbh)
}
\arguments{
\item{dbh}{vector of DBH (diameter at breast height) measurements of
competitor trees}

\item{dist}{vector of distances from focal tree to competitor trees}

\item{focal_dbh}{DBH of focal tree}
}
\value{
value of competition index for focal tree
}
\description{
Martin and Ek 1984
}
\details{
Gives the sum of ratios of competitor to focal tree DBHs, multiplied
by the exponential of distances divided by competitor DBH plus focal DBH.
Given by the equation:
\deqn{\sum_{j\neq{}i}^{n} (d_{j} / d_{i}) exp((16l_{ij}) / (d_{i} + d_{j}))}
}
\examples{
data(bicuar)
nb <- nearNeighb(bicuar$x, bicuar$y, bicuar$stem_id, k = 4) 
lapply(nb, function(x) {
  nb <- merge(x, bicuar, by.x = "nb", by.y = "stem_id")
  focal_diam <- unique(bicuar[bicuar$stem_id == nb$focal,"diam"])
  martinEk(nb$diam, nb$nb_dist, focal_diam)
  })

}
\references{
Martin G. L., Ek A. R. (1984). A Comparison of Competition
Measures and Growth Models for Predicting Plantation Red Pine Diameter and
Height Growth. Forest Science. Volume 30. Pages 731-743.
}
