% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearNeighb.R
\name{nearNeighb}
\alias{nearNeighb}
\title{Find nearest neighbours within a radius}
\usage{
nearNeighb(x, y, id = NULL, k = NULL, radius = NULL, zones = NULL)
}
\arguments{
\item{x}{vector of individual x axis coordinates}

\item{y}{vector of individual y axis coordinates}

\item{id}{vector of individual IDs. If NULL, vector positions are used.}

\item{k}{number of neighbours to search for, starting from nearest in
coordinate space. If NULL, \code{radius} must be provided.}

\item{radius}{radius to look for nearest neighbours, in units of XY
coordinates. If NULL, \code{k} must be provided.}

\item{zones}{number of zones of equal arc angle, e.g. zones == 4 results in
four zones each with 90deg arc. If NULL, no zones are defined. If zones
are defined, the nearest competitor within each zone is returned. If
zones are defined, \code{radius} must also be defined.}
}
\value{
list of dataframes per focal tree, of neighbours, their distances
and angles. If no competitors are found within the radius of a focal
tree, NA is returned for all columns except focal ID.
}
\description{
Find nearest neighbours within a radius
}
