% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearNeighb.R
\name{nearNeighb}
\alias{nearNeighb}
\title{Find nearest neighbours within a radius}
\usage{
nearNeighb(x, y = NULL, k = NULL, radius = NULL, zones = NULL)
}
\arguments{
\item{x}{two column matrix of individual x and y coordinates}

\item{y}{optional two column matrix of individual x and y coordinates}

\item{k}{number of neighbours to search for, starting from nearest in
coordinate space. If NULL, \code{radius} must be provided.}

\item{radius}{radius to look for nearest neighbours, in units of XY
coordinates. If \code{NULL}, \code{k} must be provided.}

\item{zones}{number of zones of equal arc angle, e.g. \code{zones = 4} results in
four zones each with 90deg arc. If \code{NULL}, no zones are defined. If zones
are defined, the nearest competitor within each zone is returned. If
zones are defined, \code{radius} must also be defined.}
}
\value{
List of dataframes per focal individual in x, of neighbours,
their distances and angles relative to the focal individual. If no
competitors are found within the radius of a focal individual,
NA is returned for all columns except focal ID.
}
\description{
Find nearest neighbours within a radius
}
\details{
If \code{y} is provided, nearest neighbours of individuals in
\code{y} are identified for each individual in \code{x}, otherwise,
nearest neighbours in \code{x} are identified.

In the case of ties, the first nearest neighbour is returned.
}
\examples{
data(bicuar)

nearNeighb(bicuar[,c("x", "y")], k = 4)
nearNeighb(bicuar[1:10, c("x", "y")], bicuar[, c("x", "y")], radius = 5)
nearNeighb(bicuar[,c("x", "y")], radius = 5, zones = 4)

}
