% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pielou.R
\name{pielou}
\alias{pielou}
\title{Pielou's index of non-randomness}
\usage{
pielou(x, y, xmin, xmax, ymin, ymax, k)
}
\arguments{
\item{x}{vector of individual x axis coordinates}

\item{y}{vector of individual y axis coordinates}

\item{xmin}{minimum x coordinate in plot}

\item{xmax}{maximum x coordinate in plot}

\item{ymin}{minimum y coordinate in plot}

\item{ymax}{maximum y coordinate in plot}

\item{k}{number of randomly allocated sample points}
}
\value{
Value of Pielou's index
}
\description{
Pielou's index of non-randomness
}
\details{
As the sample points are randomly allocated within the bounds of
xmin,xmax,ymin,ymax, the mean of a number of runs of this function
could be used to further constrain the estimate of Pielou's index.
}
\references{
Pielou, E. C. (1959). The use of point to plant distances in
the study of the pattern of plan populations. Journal of Ecology.
Volume 47. Pages 607-613.
}
