% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winkelmass.R
\name{winkelmass}
\alias{winkelmass}
\title{Calculate the Winkelmass (spatial regularity of trees)}
\usage{
winkelmass(x, y, k = 4)
}
\arguments{
\item{x}{vector of individual x axis coordinates}

\item{y}{vector of individual y axis coordinates}

\item{k}{number of neighbours to consider}
}
\value{

}
\description{
Calculate the Winkelmass (spatial regularity of trees)
}
\references{
von Gadow, K., Hui, G. Y. (2001). Characterising forest spatial
structure and diversity. Sustainable Forestry in Temperate Regions. Proc. of
an international workshop organized at the University of Lund, Sweden.
Pages 20- 30.
}
